Added distributed_computation from PyPyContact
distributed_computation.py is a submodule that allows to easily set up
embarrassingly parallel jobs on disparate computers or clusters. See the
examples in the distributed_computation folder for a hint about how to
use it.

distributed_computation.py is maintained in the project PyPyContact and
may be more up to date over there. If you wish to modify or extend it,
please consider doing it in PyPyContact, in order to keep the lines of
development together. Contact till.junge@kit.edu (or
till.junge@altermail.ch) for access to PyPyContact.