Added option to divide EAM Hessian by mass to calculate dynamical matrix
Division by mass is done directly during construction, in order to
avoid making a copy of the full hessian (which would be required
otherwise). A test has been added to ensure that the dynamical
matrix is symmetric, and that division during construction yields
the same result than division after construction of the Hessian.