BUG: EAM force calculation
There was an error in the EAM force calculation. The contribution
of the derivative of the embeddding energy contains a term
(derivative of embedding energy of atom a) *
(derivative of electron density of atom b) +
(derivative of embedding energy of atom b) *
(derivative of electron density of atom a).

In the previous version, the derivative of the
electron density of atom a was used two times, i.e.
(
(derivative of embedding energy of atom a) +
(derivative of embedding energy of atom b) *
) *
(derivative of electron density of atom a).

This formula is wrong if a and b are not the same type. A test was added to
ensure that the force calculation is correct for a system with two atom types.

Minor modification in io.py to ensure that potential files
are closed after reading.