Removed constrain on system size in EAM Hessian calculation.
It is now possible to calculate the EAM Hessian also for
small periodic systems, where atom pairs (i,j) appear more
than one time in the neighbor list, with different entries
corresponding to different pair distances.

The EAM Hessian requires a list of common neighbors of
atom pairs. Construction of this list is independent of
EAM; therefore I have added it as a standalone method
to matscipy.neighbors.