Refactored read_eam
The number of strings and numbers in each EAM table is well-defined
and can be calculated after reading the header info (ignoring comments
in the header and after '#'). The new version of read_eam calculates
the expected number of values and checks if the table is correct. We
raise a ValueError if this is not so. The motivation for this change
is that read_eam failed in some cases where the table contained extra
values, but it was not clear that these extra values were to blame.

eam/alloy and eam/fs file styles are very similar. The previous
version of read_eam had some repeated code. The present
version makes use of the similarities to reduce code size.

Renamed a few variables for clarity.