Make sure that l is rotated into the x direction.
The current code does the opposite: align x with l.
Suppose l=[1,0,1], then x will now be [1,0,-1] rather
than the expected [1,0,1].
In lower symmetries (starting with monoclinic) these two
directions are not equivalent.

The correctness of this patch is contingent on ELATE doing
the correct thing, because it was validated against ELATE.